Agregar excepciones personalizadas para correos y teléfonos duplicados, implementar métodos de verificación en repositorios, y añadir funcionalidad para obtener copias disponibles de libros.

+++ src/main/java/com/emmanuel/biblioteca/service/CopiaLibroService.java
@@ -2,6 +2,7 @@ package com.emmanuel.biblioteca.service;
 
 import com.emmanuel.biblioteca.entity.CopiaLibro;
 import com.emmanuel.biblioteca.entity.Libro;
+import com.emmanuel.biblioteca.exception.InvalidRequestException;
 import com.emmanuel.biblioteca.repository.CopiaLibroRepository;
 import com.emmanuel.biblioteca.repository.LibroRepository;
 import com.emmanuel.biblioteca.exception.ResourceNotFoundException;
@@ -30,8 +31,14 @@ public class CopiaLibroService {
 
     public void deleteCopiaLibroById(Integer id) {
         CopiaLibro copiaLibro = getCopiaLibroById(id); // Lanza excepción si no existe
+
+        if (!copiaLibro.isDisponible()) {
+            throw new InvalidRequestException("No se puede eliminar una copia de libro que está actualmente prestada.");
+        }
+
         copiaLibroRepository.delete(copiaLibro);
     }
+
 
     public CopiaLibro saveOrUpdateCopiaLibro(Integer libroId) {
         Libro libro = libroRepository.findById(libroId)
@@ -53,5 +60,20 @@ public class CopiaLibroService {
         } catch (Exception e) {
             throw new ResourceNotFoundException("Error al obtener libros disponibles: " + e.getMessage());
         }
+    }
+
+    public List<CopiaLibro> getCopiasDisponiblesByLibroId(Integer libroId) {
+        // Verificar si el libro existe
+        if (!libroRepository.existsById(libroId)) {
+            throw new ResourceNotFoundException("Libro con ID " + libroId + " no encontrado");
+        }
+
+        List<CopiaLibro> copiasDisponibles = copiaLibroRepository.findByLibroIdAndDisponible(libroId, true);
+
+        if (copiasDisponibles.isEmpty()) {
+            throw new ResourceNotFoundException("No hay copias disponibles para el libro con ID " + libroId);
+        }
+
+        return copiasDisponibles;
     }
 }
